# gen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package models

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type AnalyseTask struct {
	ID         int32
	Name       string
	PatientID  pgtype.Int4
	FilemetaID pgtype.Int4
	Status     int16
	Result     int16
	Predict    pgtype.Text
	CreatedAt  pgtype.Timestamptz
	UpdatedAt  pgtype.Timestamptz
}

type Filemeta struct {
	ID          int32
	Format      int16
	Size        int32
	Filename    string
	ContentType string
	Key         string
	PatientID   pgtype.Int4
	CreatedAt   pgtype.Timestamptz
	UpdatedAt   pgtype.Timestamptz
}

type Patient struct {
	ID        int32
	Name      string
	Surname   string
	Bdate     pgtype.Date
	CreatedAt pgtype.Timestamptz
	UpdatedAt pgtype.Timestamptz
}

type User struct {
	ID           pgtype.UUID
	Role         int32
	Name         string
	Login        string
	PasswordHash []byte
	CreatedAt    pgtype.Timestamptz
	UpdatedAt    pgtype.Timestamptz
}